完善了获取table表名的字符串操作，以及mybatis-plus的事务整合操作

+++ report-web/src/main/java/com/hyc/report/entity/ReportDatabase.java
@@ -25,7 +25,7 @@ public class ReportDatabase extends BaseEntity implements Serializable {
 
     @ApiModelProperty(value = "数据源配置id",example = "1")
     @TableId(value="DATABASE_ID",type = IdType.INPUT)
-    private Long databaseId;
+    private Integer databaseId;
 
     @ApiModelProperty(value = "数据源配置名称",example = "B2I数据源")
     @TableField("DATABASE_NAME")
